requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
sudachipy
sudachidict_core
sudachidict_full
requests
python-dotenv